Clean up /host
